Now with nicer rendering and better scan outputs!
